Move the metrics extractor to its own pkg under plugins/datalayer/extractor
