+++ pkg/epp/datalayer/logger/logger.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metrics
+package logger
 
 import (
 	"context"
