+++ pkg/epp/framework/plugins/datalayer/extractor/metrics/spec_test.go
@@ -27,6 +27,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
 	"k8s.io/utils/ptr"
+
+	sourcemetrics "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/plugins/datalayer/source/metrics"
 )
 
 // --- Test Helpers ---
@@ -207,7 +209,7 @@ func TestStringToMetricSpec(t *testing.T) {
 
 // TestGetMetric checks retrieving of standard metrics from a metric families.
 func TestGetMetric(t *testing.T) {
-	metricFamilies := PrometheusMetricMap{
+	metricFamilies := sourcemetrics.PrometheusMetricMap{
 		"metric1": makeMetricFamily("metric1",
 			makeMetric(map[string]string{"label1": "value1"}, 1.0, 1000),
 			makeMetric(map[string]string{"label1": "value2"}, 2.0, 2000),
@@ -333,7 +335,7 @@ func TestGetLoRAMetric(t *testing.T) {
 
 	testCases := []struct {
 		name             string
-		metricFamilies   PrometheusMetricMap
+		metricFamilies   sourcemetrics.PrometheusMetricMap
 		expectedAdapters map[string]int
 		expectedMax      int
 		expectedErr      error
@@ -341,7 +343,7 @@ func TestGetLoRAMetric(t *testing.T) {
 	}{
 		{
 			name: "no lora metrics",
-			metricFamilies: PrometheusMetricMap{
+			metricFamilies: sourcemetrics.PrometheusMetricMap{
 				"some_other_metric": makeMetricFamily("some_other_metric",
 					makeMetric(nil, 1.0, 1000),
 				),
@@ -353,7 +355,7 @@ func TestGetLoRAMetric(t *testing.T) {
 		},
 		{
 			name: "basic lora metrics",
-			metricFamilies: PrometheusMetricMap{
+			metricFamilies: sourcemetrics.PrometheusMetricMap{
 				"vllm:lora_requests_info": makeMetricFamily("vllm:lora_requests_info",
 					makeMetric(map[string]string{"running_lora_adapters": "lora1", "max_lora": "2"}, 3000.0, 1000),       // Newer
 					makeMetric(map[string]string{"running_lora_adapters": "lora2,lora3", "max_lora": "4"}, 1000.0, 1000), // Older
@@ -367,7 +369,7 @@ func TestGetLoRAMetric(t *testing.T) {
 		},
 		{
 			name: "no matching lora metrics",
-			metricFamilies: PrometheusMetricMap{
+			metricFamilies: sourcemetrics.PrometheusMetricMap{
 				"vllm:lora_requests_info": makeMetricFamily("vllm:lora_requests_info",
 					makeMetric(map[string]string{"other_label": "value"}, 5.0, 3000),
 				),
